Move generic box properties to core.
