+++ masonry/src/widgets/align.rs
@@ -18,8 +18,7 @@ use crate::core::{
 };
 use crate::core::{MeasureCtx, WidgetMut};
 use crate::kurbo::{Axis, Rect, Size};
-use crate::layout::{LayoutSize, LenReq, SizeDef};
-use crate::properties::types::UnitPoint;
+use crate::layout::{LayoutSize, LenReq, SizeDef, UnitPoint};
 use crate::util::include_screenshot;
 
 // TODO - Have child widget type as generic argument
