+++ masonry_core/src/properties/mod.rs
@@ -3,6 +3,20 @@
 
 //! Core properties.
 
+mod background;
+mod border_color;
+mod border_width;
+mod box_shadow;
+mod corner_radius;
 mod dimensions;
+mod padding;
 
+pub mod types;
+
+pub use background::*;
+pub use border_color::*;
+pub use border_width::*;
+pub use box_shadow::*;
+pub use corner_radius::*;
 pub use dimensions::*;
+pub use padding::*;
