+++ xilem/examples/http_cats.rs
@@ -8,8 +8,7 @@
 
 use std::sync::Arc;
 
-use masonry::layout::{AsUnit, Length};
-use masonry::properties::types::UnitPoint;
+use masonry::layout::{AsUnit, Length, UnitPoint};
 use masonry::properties::{LineBreaking, Padding};
 use tokio::sync::mpsc::UnboundedSender;
 use vello::peniko::{Blob, ImageAlphaType, ImageData, ImageFormat};
